refactor: move mnemonic messages from accounts handler
