+++ apps/extension/src/ui/domains/Account/AccountAdd/AccountAddSecret/AccountAddSecretMnemonicForm.tsx
@@ -86,7 +86,7 @@ const testNoDuplicate = async (
 const testValidMnemonic = async (val?: string) => {
   // Don't bother calling the api if the mnemonic isn't the right length to reduce Sentry noise
   if (!val || ![12, 24].includes(val.split(" ").length)) return false
-  return await api.accountValidateMnemonic(val)
+  return await api.validateMnemonic(val)
 }
 
 export const AccountAddSecretMnemonicForm = () => {
